requirements.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.30.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.8.6
wsproto==1.2.0
zipp==3.15.0
